requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
astrapy==1.5.0
async-timeout==4.0.3
attrs==24.2.0
cassandra-driver==3.29.2
cassio==0.1.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
frozenlist==1.4.1
geomet==0.2.1.post1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-astradb==0.4.0
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langchainhub==0.1.21
langsmith==0.1.125
multidict==6.1.0
numpy==1.26.4
openai==1.47.0
orjson==3.10.7
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
uuid6==2024.7.10
yarl==1.11.1
